Update signiture abbreviations
The abbreviations whould sometimes get in the way. Like db. I was unable
to type db anywhere. Remember that vim abbreviations work in the vim
command prompt too. This is the problem with db for a signiture and
rails rake tasks that begin with db.

Amos King @adkron <amos.l.king@gmail.com>